requirements.txt
Augmentor==0.2.8
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.16.0
torch==1.8.1
tqdm==4.60.0
typing-extensions==3.10.0.0
